add benchmark of integer partition comparing dynamic programming and constraint programming

+++ algorithm/src/jmh/java/dev/carbonshow/algorithm/partition/MaxPartitionBenchmark.java
@@ -10,12 +10,12 @@ import java.util.concurrent.TimeUnit;
 @State(Scope.Benchmark)
 public class MaxPartitionBenchmark {
 
-    final private Map<Integer, Integer> ADDENDS = Map.of(1, 100, 2, 40, 5, 10);
-    final private int PARTITIONED = 10;
-    final private ArrayList<Map<Integer, Long>> PARTITION_PLANS = new ArrayList<>(Arrays.asList(Map.of(1, 6L, 2, 2L),
-            Map.of(1, 1L, 2, 2L, 5, 1L),
-            Map.of(5, 2L),
-            Map.of(1, 5L, 5, 1L)
+    final private Map<Long, Long> ADDENDS = Map.of(1L, 100L, 2L, 40L, 5L, 10L);
+    final private long PARTITIONED = 10;
+    final private ArrayList<Map<Long, Long>> PARTITION_PLANS = new ArrayList<>(Arrays.asList(Map.of(1L, 6L, 2L, 2L),
+            Map.of(1L, 1L, 2L, 2L, 5L, 1L),
+            Map.of(5L, 2L),
+            Map.of(1L, 5L, 5L, 1L)
     ));
 
     final private DefaultMaxPartitions solver = new DefaultMaxPartitions();
