+++ algorithm/src/main/java/dev/carbonshow/algorithm/partition/MaxPartitions.java
@@ -41,5 +41,5 @@ interface MaxPartitions {
    * @param partitioned   对上面的加数集合进行划分，每个划分所含加数之和应该等于该值。
    * @return 返回最终划分结果的详细数据，将所有划分方案的组成以及实施数量罗列出来；不存在则返回 null
    */
-  ArrayList<PartitionData> solve(Map<Integer, Integer> addends, int partitioned);
+  ArrayList<PartitionData> solve(Map<Long, Long> addends, long partitioned);
 }
